use tap instead of ava, increase test coverage to 100 %
